feat(test): introduce test.config.js
This file encapsulate everything specific to playwright testing. Now
the general test.js can be reused for other projects.
